我不是機器人 完成

+++ src/app/components/Member/resetpassword/resetpassword.component.ts
@@ -5,6 +5,10 @@ import { LoginService } from 'src/app/service/Member/login.service';
 import { MatDialog } from '@angular/material/dialog';
 import { ResetPasswordEmailComponent } from '../reset-password-email/reset-password-email.component';
 import { ResetPasswordCompleteComponent } from '../reset-password-complete/reset-password-complete.component';
+import Swal from 'sweetalert2';
+
+//我不是機器人相關
+declare var grecaptcha: any;
 
 @Component({
   selector: 'app-resetpassword',
@@ -15,6 +19,10 @@ export class ResetpasswordComponent {
 
   constructor(private loginService: LoginService, private router: Router,private dialog: MatDialog ) { }
 
+    //我不是機器人相關
+    captchaResponse: string | null = null;
+    recaptchaRendered = false;
+    //===================================================
 
   forgetPasswordTransfer: forgetPasswordTransfer =
     {
@@ -58,15 +66,49 @@ export class ResetpasswordComponent {
         this.errormessage.passworcomfirm = '密碼不一致';
         return;
       }
-    this.loginService.ResetPasswordAPI(this.forgetPasswordTransfer).subscribe(data => {
-    if(data.result ==='success')
+
+    //打API
+    if (this.captchaResponse)
       {
-        this.dialog.open(ResetPasswordCompleteComponent);
-        this.loginService.removeForgetPasswordEmail();
-        this.router.navigate(['**']);
+        this.loginService.ResetPasswordAPI(this.forgetPasswordTransfer).subscribe(data => {
+          if(data.result ==='success')
+            {
+              this.dialog.open(ResetPasswordCompleteComponent);
+              this.loginService.removeForgetPasswordEmail();
+              this.router.navigate(['**']);
+            }
+          })
       }
-    })
+      else
+      {
+        Swal.fire({
+          icon: "error",
+          title: "請完成我不是機器人驗證",
+          showConfirmButton: false,
+        })
+      }
+
   }
-
-
+    // 初始化 reCAPTCHA
+    ngAfterViewChecked(): void {
+      if (!this.recaptchaRendered) {
+        grecaptcha.render('recaptcha-container', {
+          sitekey: '6Le6oHoqAAAAAPL4kjsNmc3Uyd9WIadivdAKzCnR',
+          callback: (response: string) => this.onCaptchaResolved(response),
+        });
+        this.recaptchaRendered = true;
+      }
+    }
+      // 初始化 reCAPTCHA
+      ngAfterViewInit(): void {
+        grecaptcha.render('recaptcha-container', {
+          sitekey: '6Le6oHoqAAAAAPL4kjsNmc3Uyd9WIadivdAKzCnR', // 使用你的 Site Key
+          callback: (response: string) => this.onCaptchaResolved(response),
+        });
+      }
+    // 當 reCAPTCHA 被解決後，回調此函數
+    onCaptchaResolved(captchaResponse: string) {
+      console.log('reCAPTCHA 回應:', captchaResponse);
+      this.captchaResponse = captchaResponse;
+    }
 }
